requirements.txt
Django==6.0.4
whitenoise>=6.0
python-dotenv>=1.0
gunicorn>=20.1.0
psycopg2-binary>=2.9
django-cors-headers>=4.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.5.1
drf-spectacular>=0.26.0
